ApeSwap UniswapV2 & MasterChef events
This PR adds events that cover the UniswapV2 and MasterChef components
of ApeSwap. Unlike many projects the contract names were changed from
e.g. UniswapV2Factory to ApeFactory and from MasterChef to MasterApe.

This change means there is some shim logic to use these events in
downstream ETLs that can take Uniswap/Sushi shaped inputs. It is
presumably better to respect the contract names than to ignore it for
the sake of downstream compatibility. But thoughts on precedent
are welcome here.