Fix table definitions for Ronin Axie events (#80)
The BigQuery UDF we use to parse events recursively processes tuples, and turns
them into Javascript Objects. If the target column in BigQuery is defined as a
STRING, the UDF ultimately calls .toString() on this object, which returns
"[object Object]" instead of a useful representation. We can fix this by providing
a full schema for these objects.